Allow textdomain to be omitted, but give error
Previously we were inconsistently requiring the text domain for l10n
functions. Now it’s not required for any of them, but if you leave it
out, it will be reported as an unexpected domain `’’`, unless you add
that to the list of valid domains in your config file.

Fixes #7