Avoid cache bloat
Over time the cache would bloat. This fixes that by starting from
scratch for the parsed files, while keeping the ignores cache untouched
for the unchanged files.

It should also fixes an issue where the prior ignores in a file might
be dropped from the cache when new ignores were being added to the file.

Fixes #21