Use correct function when checking if an arg is ignored
Sometimes I would get unexpected “not gettexted” errors for function
arguments that were supposed to be ignored. The cause of the bug was
that sometimes the incorrect function name would be used to check if
the arg was ignored. The error would contain the correct function name,
because the `prepare_gettext()` method exits any “functions” whose
arguments haven’t started yet. It is called before the message is
generated, but, critically, _after_ the check for whether the function
args were ignored. To fix this, the check has been moved inside of the
method, after the logic that exits unstarted “functions”.

Some basic logic remains at the original location, only because it
helps us short-circuit the check without having to call a function.

Unit tests included.

Fixes #28