Allow class properties to be ignored
In the process I’ve moved some logic out of the `prepare_non_gettext()`
method that really didn’t need to be inside of it. It is only needed
for one of the three places that the function is used, so I’ve moved
the logic to that location instead.

Fixes #26